+++ test/passing/partial.ml
@@ -1,0 +1,20 @@
+let foooooo
+= fooooooooo
+
+let foooooooo
+= bar baaaaar barrr
+
+module K = struct
+  let k
+
+;;
+begin
+  let x = in
+  ()
+
+;;
+let foooooo =
+  fooooo
+  foooooooo
+
+let k =